CharacterSetup struct
